Move the OpenGL specific Skia surface handling into a separate module
